+++ src/field_specials.c
@@ -2312,7 +2312,8 @@ void ShowScrollableMultichoice(void)
         break;
     case SCROLL_MULTI_BF_EXCHANGE_CORNER_VITAMIN_VENDOR:
         task->tMaxItemsOnScreen = MAX_SCROLL_MULTI_ON_SCREEN;
-        task->tNumItems = 12;
+        // NOTE: I tried to change numItems to not be hardcoded, but the game broke when I did
+        task->tNumItems = 11;
         task->tLeft = 14;
         task->tTop = 1;
         task->tWidth = 15;
@@ -2458,7 +2459,6 @@ static const u8 *const sScrollableMultichoiceOptions[][MAX_SCROLL_MULTI_LENGTH] 
         gText_HpUp1BP,
         gText_PPUp1BP,
         gText_RareCandy8BP,
-        gText_ExpShare8BP,
         gText_MachoBrace8BP,
         gText_MasterBall256BP,
         gText_Exit