+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Re:Emerald Changelog
 
 ## In-development
+**Update 14**
+- Removed the Exp. Share from the Battle Frontier exchange, since it's now a Key Item
+
 **Update 13**
 - Added the Medicine and Battle Items pockets to the bag
 - Added a Potion to the bag when starting a new game, to replace the one that used to be in the PC storage